# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kaai Code Snippet API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SnippetStatus = (typeof SnippetStatus)[keyof typeof SnippetStatus];

export const SnippetStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Snippet {
  id: string;
  title: string;
  description: string;
  language: string;
  tags: string[];
  code: string;
  authorName: string;
  authorEmail: string;
  status: SnippetStatus;
  rejectReason?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface SnippetListResponse {
  data: Snippet[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateSnippetBody {
  /**
   * @minLength 1
   * @maxLength 200
   */
  title: string;
  /**
   * @minLength 1
   * @maxLength 1000
   */
  description: string;
  language: string;
  /** @minItems 1 */
  tags: string[];
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  authorName: string;
  authorEmail: string;
}

export interface Stats {
  totalSnippets: number;
  pendingSnippets: number;
  approvedSnippets: number;
  rejectedSnippets: number;
  totalAuthors: number;
  totalLanguages: number;
}

export interface LanguageStat {
  language: string;
  count: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ListSnippetsParams = {
  search?: string;
  language?: string;
  tag?: string;
  sortBy?: "popular" | "latest" | "az";
  page?: number;
  limit?: number;
};

export type RejectSnippetBody = {
  reason?: string;
};

export type GetRecentSnippetsParams = {
  limit?: number;
};