# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kaai Code Snippet API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List approved snippets
 */
export const listSnippetsQueryPageDefault = 1;
export const listSnippetsQueryLimitDefault = 12;

export const ListSnippetsQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  language: zod.coerce.string().optional(),
  tag: zod.coerce.string().optional(),
  sortBy: zod.enum(["popular", "latest", "az"]).optional().default("popular"),
  page: zod.coerce.number().default(listSnippetsQueryPageDefault),
  limit: zod.coerce.number().default(listSnippetsQueryLimitDefault),
});

export const ListSnippetsResponse = zod.object({
  data: zod.array(
    zod.object({
      id: zod.string(),
      title: zod.string(),
      description: zod.string(),
      language: zod.string(),
      tags: zod.array(zod.string()),
      code: zod.string(),
      authorName: zod.string(),
      authorEmail: zod.string(),
      status: zod.enum(["pending", "approved", "rejected"]),
      rejectReason: zod.string().nullish(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
  totalPages: zod.number(),
});

/**
 * @summary Upload a new snippet
 */
export const createSnippetBodyTitleMax = 200;

export const createSnippetBodyDescriptionMax = 1000;

export const CreateSnippetBody = zod.object({
  title: zod.string().min(1).max(createSnippetBodyTitleMax),
  description: zod.string().min(1).max(createSnippetBodyDescriptionMax),
  language: zod.string(),
  tags: zod.array(zod.string()).min(1),
  code: zod.string().min(1),
  authorName: zod.string().min(1),
  authorEmail: zod.string().email(),
});

/**
 * @summary Get snippet by ID
 */
export const GetSnippetParams = zod.object({
  id: zod.coerce.string(),
});

export const GetSnippetResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  description: zod.string(),
  language: zod.string(),
  tags: zod.array(zod.string()),
  code: zod.string(),
  authorName: zod.string(),
  authorEmail: zod.string(),
  status: zod.enum(["pending", "approved", "rejected"]),
  rejectReason: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Approve a pending snippet (bot only)
 */
export const ApproveSnippetParams = zod.object({
  id: zod.coerce.string(),
});

export const ApproveSnippetResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  description: zod.string(),
  language: zod.string(),
  tags: zod.array(zod.string()),
  code: zod.string(),
  authorName: zod.string(),
  authorEmail: zod.string(),
  status: zod.enum(["pending", "approved", "rejected"]),
  rejectReason: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Reject a pending snippet (bot only)
 */
export const RejectSnippetParams = zod.object({
  id: zod.coerce.string(),
});

export const RejectSnippetBody = zod.object({
  reason: zod.string().optional(),
});

export const RejectSnippetResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  description: zod.string(),
  language: zod.string(),
  tags: zod.array(zod.string()),
  code: zod.string(),
  authorName: zod.string(),
  authorEmail: zod.string(),
  status: zod.enum(["pending", "approved", "rejected"]),
  rejectReason: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary List pending snippets (admin)
 */
export const ListPendingSnippetsResponse = zod.object({
  data: zod.array(
    zod.object({
      id: zod.string(),
      title: zod.string(),
      description: zod.string(),
      language: zod.string(),
      tags: zod.array(zod.string()),
      code: zod.string(),
      authorName: zod.string(),
      authorEmail: zod.string(),
      status: zod.enum(["pending", "approved", "rejected"]),
      rejectReason: zod.string().nullish(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
  totalPages: zod.number(),
});

/**
 * @summary Get platform statistics
 */
export const GetStatsResponse = zod.object({
  totalSnippets: zod.number(),
  pendingSnippets: zod.number(),
  approvedSnippets: zod.number(),
  rejectedSnippets: zod.number(),
  totalAuthors: zod.number(),
  totalLanguages: zod.number(),
});

/**
 * @summary Get snippet counts by language
 */
export const GetLanguageStatsResponseItem = zod.object({
  language: zod.string(),
  count: zod.number(),
});
export const GetLanguageStatsResponse = zod.array(GetLanguageStatsResponseItem);

/**
 * @summary Get recently approved snippets
 */
export const getRecentSnippetsQueryLimitDefault = 5;

export const GetRecentSnippetsQueryParams = zod.object({
  limit: zod.coerce.number().default(getRecentSnippetsQueryLimitDefault),
});

export const GetRecentSnippetsResponseItem = zod.object({
  id: zod.string(),
  title: zod.string(),
  description: zod.string(),
  language: zod.string(),
  tags: zod.array(zod.string()),
  code: zod.string(),
  authorName: zod.string(),
  authorEmail: zod.string(),
  status: zod.enum(["pending", "approved", "rejected"]),
  rejectReason: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetRecentSnippetsResponse = zod.array(
  GetRecentSnippetsResponseItem,
);