# lib/api-zod/src/generated/types/createSnippetBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kaai Code Snippet API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSnippetBody {
  /**
   * @minLength 1
   * @maxLength 200
   */
  title: string;
  /**
   * @minLength 1
   * @maxLength 1000
   */
  description: string;
  language: string;
  /** @minItems 1 */
  tags: string[];
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  authorName: string;
  authorEmail: string;
}